reorganize files to encompass display files
